+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Wearable Artifacts can now be enchanted with Curse of Binding
+- Support for myloot chest textures on mimics
+
 ### Changed
 - Reset configs when they fail to load, rather than crashing (#26)